refactor: improve delegates error handling
